+++ server/src/test/java/org/realityforge/replicant/server/ChangeAccumulatorTest.java
@@ -1,10 +1,11 @@
 package org.realityforge.replicant.server;
 
 import java.io.Serializable;
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.Map;
 import javax.json.Json;
 import javax.json.JsonObject;
+import org.realityforge.guiceyloops.shared.ValueUtil;
 import org.realityforge.replicant.server.ChannelAction.Action;
 import org.realityforge.replicant.server.transport.Packet;
 import org.realityforge.replicant.server.transport.ReplicantSession;
@@ -28,7 +29,8 @@ public class ChangeAccumulatorTest
     final String subChannelId = "2";
 
     accumulator.addChange( c, new Change( message, channelId, subChannelId ) );
-    final boolean impactsInitiator = accumulator.complete( "s1", "j1" );
+    final Integer requestId = ValueUtil.randomInt();
+    final boolean impactsInitiator = accumulator.complete( "s1", requestId );
 
     assertTrue( impactsInitiator );
     assertEquals( c.getQueue().size(), 1 );
@@ -37,7 +39,7 @@ public class ChangeAccumulatorTest
     assertEquals( change.getId(), "42#myID" );
     assertEquals( change.getEntityMessage().getId(), id );
     assertEquals( change.getEntityMessage().getTypeId(), typeID );
-    assertEquals( packet.getRequestId(), "j1" );
+    assertEquals( packet.getRequestId(), requestId );
     final Map<Integer, Serializable> channels = change.getChannels();
     assertEquals( channels.size(), 1 );
     assertEquals( channels.get( channelId ), subChannelId );
@@ -57,11 +59,12 @@ public class ChangeAccumulatorTest
 
     final EntityMessage message = MessageTestUtil.createMessage( id, typeID, 0, "r1", "r2", "a1", "a2" );
 
-    accumulator.addChanges( c, Arrays.asList( new Change( message, 1, 0 ) ) );
+    accumulator.addChanges( c, Collections.singletonList( new Change( message, 1, 0 ) ) );
 
     assertEquals( accumulator.getChangeSet( c ).getChanges().size(), 1 );
 
-    final boolean impactsInitiator = accumulator.complete( "s1", "j1" );
+    final Integer requestId = ValueUtil.randomInt();
+    final boolean impactsInitiator = accumulator.complete( "s1", requestId );
 
     assertEquals( accumulator.getChangeSet( c ).getChanges().size(), 0 );
 
@@ -69,7 +72,7 @@ public class ChangeAccumulatorTest
     assertEquals( c.getQueue().size(), 1 );
     final Packet packet = c.getQueue().nextPacketToProcess();
     assertEquals( packet.getChangeSet().getChanges().iterator().next().getEntityMessage().getId(), id );
-    assertEquals( packet.getRequestId(), "j1" );
+    assertEquals( packet.getRequestId(), requestId );
 
     accumulator.complete( null, null );
     assertEquals( c.getQueue().size(), 1 );
@@ -83,11 +86,14 @@ public class ChangeAccumulatorTest
 
     final JsonObject filter = Json.createBuilderFactory( null ).createObjectBuilder().build();
     accumulator.addActions( c,
-                            Arrays.asList( new ChannelAction( new ChannelAddress( 1, 2 ), Action.ADD, filter ) ) );
+                            Collections.singletonList( new ChannelAction( new ChannelAddress( 1, 2 ),
+                                                                          Action.ADD,
+                                                                          filter ) ) );
 
     assertEquals( accumulator.getChangeSet( c ).getChannelActions().size(), 1 );
 
-    final boolean impactsInitiator = accumulator.complete( "s1", "j1" );
+    final Integer requestId = ValueUtil.randomInt();
+    final boolean impactsInitiator = accumulator.complete( "s1", requestId );
 
     assertEquals( accumulator.getChangeSet( c ).getChannelActions().size(), 0 );
 
@@ -98,7 +104,7 @@ public class ChangeAccumulatorTest
     assertEquals( action.getChannelAddress().getChannelId(), 1 );
     assertEquals( action.getAction(), Action.ADD );
     assertEquals( action.getFilter(), filter );
-    assertEquals( packet.getRequestId(), "j1" );
+    assertEquals( packet.getRequestId(), requestId );
 
     assertEquals( c.getQueue().size(), 1 );
     accumulator.complete( null, null );
@@ -114,7 +120,8 @@ public class ChangeAccumulatorTest
     final EntityMessage message = MessageTestUtil.createMessage( "myID", 42, 0, "r1", "r2", "a1", "a2" );
 
     accumulator.addChange( c, new Change( message, 1, 0 ) );
-    final boolean impactsInitiator = accumulator.complete( "s2", "j1" );
+    final Integer requestId = ValueUtil.randomInt();
+    final boolean impactsInitiator = accumulator.complete( "s2", requestId );
 
     assertFalse( impactsInitiator );
 
@@ -129,7 +136,8 @@ public class ChangeAccumulatorTest
     final ChangeAccumulator accumulator = new ChangeAccumulator();
 
     accumulator.getChangeSet( c );
-    final boolean impactsInitiator = accumulator.complete( "s1", "j1" );
+    final Integer requestId = ValueUtil.randomInt();
+    final boolean impactsInitiator = accumulator.complete( "s1", requestId );
 
     assertFalse( impactsInitiator );