+++ server/src/test/java/org/realityforge/replicant/server/ee/ReplicationInterceptorTest.java
@@ -79,7 +79,7 @@ public class ReplicationInterceptorTest
 
     when( em.isOpen() ).thenReturn( true );
     ReplicantContextHolder.put( ServerConstants.SESSION_ID_KEY, "s1" );
-    ReplicantContextHolder.put( ServerConstants.REQUEST_ID_KEY, "r1" );
+    ReplicantContextHolder.put( ServerConstants.REQUEST_ID_KEY, 37 );
     final Object result = interceptor.businessIntercept( context );
     verify( em ).flush();
 
@@ -89,7 +89,7 @@ public class ReplicationInterceptorTest
     assertTrue( context.isInvoked() );
     assertEquals( result, TestInvocationContext.RESULT );
     assertEquals( interceptor._sessionID, "s1" );
-    assertEquals( interceptor._requestId, "r1" );
+    assertEquals( interceptor._requestId, (Integer) 37 );
     assertNotNull( interceptor._messages );
     assertTrue( interceptor._messages.contains( message ) );
     assertEquals( ReplicantContextHolder.get( ServerConstants.REQUEST_COMPLETE_KEY ), "0" );
@@ -107,7 +107,7 @@ public class ReplicationInterceptorTest
       createInterceptor( registry, entityManager );
 
     ReplicantContextHolder.put( ServerConstants.SESSION_ID_KEY, "s1" );
-    ReplicantContextHolder.put( ServerConstants.REQUEST_ID_KEY, "r1" );
+    ReplicantContextHolder.put( ServerConstants.REQUEST_ID_KEY, 37 );
 
     when( entityManager.isOpen() ).thenReturn( true );
 
@@ -132,7 +132,7 @@ public class ReplicationInterceptorTest
 
     when( em.isOpen() ).thenReturn( true );
     ReplicantContextHolder.put( ServerConstants.SESSION_ID_KEY, "s1" );
-    ReplicantContextHolder.put( ServerConstants.REQUEST_ID_KEY, "r1" );
+    ReplicantContextHolder.put( ServerConstants.REQUEST_ID_KEY, 37 );
     final Object result = interceptor.businessIntercept( context );
     verify( em ).flush();
 
@@ -142,7 +142,7 @@ public class ReplicationInterceptorTest
     assertTrue( context.isInvoked() );
     assertEquals( result, TestInvocationContext.RESULT );
     assertEquals( interceptor._sessionID, "s1" );
-    assertEquals( interceptor._requestId, "r1" );
+    assertEquals( interceptor._requestId, (Integer) 37 );
     assertNotNull( interceptor._messages );
     assertEquals( interceptor._changeSet.getChanges().size(), 1 );
     final Change change = interceptor._changeSet.getChanges().iterator().next();
@@ -201,7 +201,7 @@ public class ReplicationInterceptorTest
 
     when( em.isOpen() ).thenReturn( true );
     ReplicantContextHolder.put( ServerConstants.SESSION_ID_KEY, "s1" );
-    ReplicantContextHolder.put( ServerConstants.REQUEST_ID_KEY, "r1" );
+    ReplicantContextHolder.put( ServerConstants.REQUEST_ID_KEY, 37 );
     try
     {
       interceptor.businessIntercept( context );
@@ -340,7 +340,7 @@ public class ReplicationInterceptorTest
     implements EntityMessageEndpoint
   {
     private String _sessionID;
-    private String _requestId;
+    private Integer _requestId;
     private Collection<EntityMessage> _messages;
     private EntityManager _entityManager;
     private final TransactionSynchronizationRegistry _registry;
@@ -365,7 +365,7 @@ public class ReplicationInterceptorTest
 
     @Override
     public boolean saveEntityMessages( @Nullable final String sessionID,
-                                       @Nullable final String requestId,
+                                       @Nullable final Integer requestId,
                                        @Nonnull final Collection<EntityMessage> messages,
                                        @Nullable final ChangeSet changeSet )
     {